pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "raidoc"
version = "1.0.6.0"
description = "Documentation for RAIMAD"
authors = [
	{name = "maybetree", email = "maybetree48@proton.me"}
	]
license = { text = "I haven't decided yet" }
readme = "README.md"
classifiers = [
	'Programming Language :: Python :: 3',
	]
requires-python = ">=3"
dependencies = [
	'jinja2',
	'libsass',
	'marko[codehilite]',
	]

#[tool.setuptools.packages.find]
#include = [ "pycif", "pycif.*" ]

[tool.ruff.lint]
exclude = [
    ]
select = [
    "E",
    "F",
    "D",
    ]
ignore = [
    "D401",  # don't complain about "imperative mood"
    ]
[tool.ruff.lint.pydocstyle]
convention = "numpy"
